pyproject.toml
[tool.poetry]
name = "kameshsampath.demos"
version = "0.0.1"
description = "Ansible Collections for OpenShift Demos"
authors = ["Kamesh Sampath <kamesh.sampath@hotmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.6"
openshift = "^0.12.0"
ansible-base = "^2.10.8"
boto3 = "^1.17.53"
pyOpenSSL = "^20.0.1"

[tool.poetry.dev-dependencies]
ansible-lint = "^5.0.7"
yamllint = "^1.26.1"
pytest = "^6.2.3"
ansible-doc-extractor = "^0.1.6"
sphinx-rtd-theme = "^0.5.2"
Sphinx = "^3.5.4"
myst-parser = "^0.13.6"
ansible-builder = "^0.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
